refactor: pages 폴더 하위에 있던 컴포넌트를 /components/teacher/threadDetail로 이동
