+++ src/components/teacher/threadDetail/ChatHeader.tsx
