+++ src/components/teacher/threadDetail/ChatInput.tsx
