+++ src/pages/teacher/threadDetail/TeacherThreadDetailPage.tsx
@@ -3,9 +3,9 @@ import { useTheme } from '@emotion/react';
 import { useMemo } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import { useThreadDetail } from '@/features/teacher/threadDetail/hooks/useThreadDetail';
-import { ChatHeader } from '@/pages/teacher/threadDetail/components/ChatHeader';
-import { ChatInput } from '@/pages/teacher/threadDetail/components/ChatInput';
-import { ChatMessageList } from '@/pages/teacher/threadDetail/components/ChatMessageList';
+import { ChatHeader } from '@/components/teacher/threadDetail/ChatHeader';
+import { ChatInput } from '@/components/teacher/threadDetail/ChatInput';
+import { ChatMessageList } from '@/components/teacher/threadDetail/ChatMessageList';
 import { InlineButton } from '@/components/common/InlineButton';
 import { Alert } from '@/components/common/Alert';
 import { Dialog } from '@/components/common/Dialog';
